11.ML/01.module.01/requirements.txt
torch
matplotlib
numpy
python-docx
